requirements.txt
python-dotenv==1.0.0
openpyxl==3.1.2
# pandas==2.1.4
pandas==2.2.3
xlsxwriter==3.1.9
gunicorn==21.2.0
flask==3.0.0
flask-sqlalchemy==3.1.1
pymssql==2.2.11
sqlalchemy==2.0.25
reportlab
flask_wtf
